feat: Introduced a runtime manager to centralize creation and disposal of PythonEngine and make exception handling easier
