+++ HelloWorldConsole/Serialization1.cs
@@ -14,22 +14,21 @@ internal static class Serialization1
 {
     public static void Run()
     {
-        Runtime.PythonDLL = @"C:\Program Files\Python312\python312.dll";
-        PythonEngine.Initialize();
+        using(var runtimeManager = new PythonRuntimeManager())
+        {            
+            var basePath = Directory.GetParent(AppDomain.CurrentDomain.BaseDirectory);
+            var scriptFilePath = Path.Combine(basePath.FullName, "PythonScripts", "serialization1.py");
 
-        var basePath = Directory.GetParent(AppDomain.CurrentDomain.BaseDirectory);
-        var scriptFilePath = Path.Combine(basePath.FullName, "PythonScripts","serialization1.py");
+            var scriptContent = File.ReadAllText(scriptFilePath, System.Text.Encoding.UTF8);
 
-        var scriptContent = File.ReadAllText(scriptFilePath, System.Text.Encoding.UTF8);
+            using(Py.GIL())
+            {
+                using var scope = Py.CreateScope();
+                scope.Exec(scriptContent);
+            }
 
-        using(Py.GIL())
-        {
-            using var scope = Py.CreateScope();
-            scope.Exec(scriptContent);
+            Console.WriteLine("Finished execution of Run method of Serialization1 class.");
         }
-
-        PythonEngine.Shutdown();
-        Console.WriteLine("Finished execution of Run method of Serialization1 class.");
     }
 }
 #endif
