+++ HelloWorldConsole/Serialization2.cs
@@ -15,22 +15,21 @@ internal static class Serialization2
 {
     public static void Run()
     {
-        Runtime.PythonDLL = @"C:\Program Files\Python312\python312.dll";
-        PythonEngine.Initialize();
-
-        var basePath = Directory.GetParent(AppDomain.CurrentDomain.BaseDirectory);
-        var scriptFilePath = Path.Combine(basePath.FullName, "PythonScripts", "serialization2.py");
-
-        var scriptContent = File.ReadAllText(scriptFilePath, System.Text.Encoding.UTF8);
-
-        using(Py.GIL())
+        using(var runtimeManager = new PythonRuntimeManager())
         {
-            using var scope = Py.CreateScope();
-            scope.Exec(scriptContent);
-        }
+            var basePath = Directory.GetParent(AppDomain.CurrentDomain.BaseDirectory);
+            var scriptFilePath = Path.Combine(basePath.FullName, "PythonScripts", "serialization2.py");
 
-        PythonEngine.Shutdown();
-        Console.WriteLine("Finished execution of Run method of Serialization2 class.");
+            var scriptContent = File.ReadAllText(scriptFilePath, System.Text.Encoding.UTF8);
+
+            using(Py.GIL())
+            {
+                using var scope = Py.CreateScope();
+                scope.Exec(scriptContent);
+            }
+
+            Console.WriteLine("Finished execution of Run method of Serialization2 class."); 
+        }
     }
 }
 #endif
