+++ HelloWorldConsole/Serialization3.cs
@@ -17,18 +17,18 @@ internal static class Serialization3
     {
         try
         {
-            Runtime.PythonDLL = @"C:\Program Files\Python312\python312.dll";
-            PythonEngine.Initialize();
-
-            var basePath = Directory.GetParent(AppDomain.CurrentDomain.BaseDirectory);
-            var scriptFilePath = Path.Combine(basePath.FullName, "PythonScripts", "serialization3.py");
-
-            var scriptContent = File.ReadAllText(scriptFilePath, System.Text.Encoding.UTF8);
-
-            using(Py.GIL())
+            using(var runtimeManager = new PythonRuntimeManager())
             {
-                using var scope = Py.CreateScope();
-                scope.Exec(scriptContent);
+                var basePath = Directory.GetParent(AppDomain.CurrentDomain.BaseDirectory);
+                var scriptFilePath = Path.Combine(basePath.FullName, "PythonScripts", "serialization3.py");
+
+                var scriptContent = File.ReadAllText(scriptFilePath, System.Text.Encoding.UTF8);
+
+                using(Py.GIL())
+                {
+                    using var scope = Py.CreateScope();
+                    scope.Exec(scriptContent);
+                } 
             }
         }
         catch(Exception ex)
@@ -39,8 +39,7 @@ internal static class Serialization3
             Console.WriteLine();
         }
         finally
-        {
-            PythonEngine.Shutdown();
+        {            
             Console.WriteLine("Finished execution of Run method of Serialization3 class.");
         }
     }
