+++ HelloWorldConsole/Serialization4.cs
@@ -11,24 +11,24 @@ using Python.Runtime;
 
 namespace HelloWorldConsole;
 
-internal class Serialization4 : IDisposable
+internal class Serialization4
 {
     public void Run()
     {
         try
         {
-            Runtime.PythonDLL = @"C:\Program Files\Python312\python312.dll";
-            PythonEngine.Initialize();
-
-            var basePath = Directory.GetParent(AppDomain.CurrentDomain.BaseDirectory);
-            var scriptFilePath = Path.Combine(basePath.FullName, "PythonScripts", "serialization4.py");
-
-            var scriptContent = File.ReadAllText(scriptFilePath, System.Text.Encoding.UTF8);
-
-            using(Py.GIL()) 
+            using(var runtimeManager = new PythonRuntimeManager())
             {
-                using var scope = Py.CreateScope();
-                scope.Exec(scriptContent);
+                var basePath = Directory.GetParent(AppDomain.CurrentDomain.BaseDirectory);
+                var scriptFilePath = Path.Combine(basePath.FullName, "PythonScripts", "serialization4.py");
+
+                var scriptContent = File.ReadAllText(scriptFilePath, System.Text.Encoding.UTF8);
+
+                using(Py.GIL())
+                {
+                    using var scope = Py.CreateScope();
+                    scope.Exec(scriptContent);
+                } 
             }
         }
         catch(Exception ex)
@@ -40,19 +40,6 @@ internal class Serialization4 : IDisposable
         }
 
         Console.WriteLine("Finished execution of Run method of Serialization4 class.");
-    }
-
-    public void Dispose()
-    {
-        Dispose(true);
-        GC.SuppressFinalize(this);
-    }
-
-    protected virtual void Dispose(bool disposing)
-    {
-        // Cleanup
-        PythonEngine.Shutdown();
-        Console.WriteLine("Dispose method of Serialization4 class was called");
     }
 }
 #endif
