+++ HelloWorldConsole/Serialization6.cs
@@ -15,22 +15,31 @@ internal static class Serialization6
 {
     public static void Run()
     {
-        Runtime.PythonDLL = @"C:\Program Files\Python312\python312.dll";
-        PythonEngine.Initialize();
-
-        var basePath = Directory.GetParent(AppDomain.CurrentDomain.BaseDirectory);
-        var scriptFilePath = Path.Combine(basePath.FullName, "PythonScripts", "serialization6.py");
-
-        var scriptContent = File.ReadAllText(scriptFilePath, System.Text.Encoding.UTF8);
-
-        using(Py.GIL())
+        try
         {
-            using var scope = Py.CreateScope();
-            scope.Exec(scriptContent);
-        }
+            using(var runtimeManager = new PythonRuntimeManager())
+            {
+                var basePath = Directory.GetParent(AppDomain.CurrentDomain.BaseDirectory);
+                var scriptFilePath = Path.Combine(basePath.FullName, "PythonScripts", "serialization6.py");
 
-        PythonEngine.Shutdown();
-        Console.WriteLine("Finished execution of Run method of Serialization6 class.");
+                var scriptContent = File.ReadAllText(scriptFilePath, System.Text.Encoding.UTF8);
+
+                using(Py.GIL())
+                {
+                    using var scope = Py.CreateScope();
+                    scope.Exec(scriptContent);
+                }
+
+                Console.WriteLine("Finished execution of Run method of Serialization6 class.");
+            }
+        }
+        catch(Exception ex)
+        {
+            Console.WriteLine();
+            Console.WriteLine(ex.Message);
+            Console.WriteLine(ex.StackTrace);
+            Console.WriteLine();
+        }
     }
 }
 #endif
